+++ src/hooks/use-tag-user.ts
@@ -1,7 +1,6 @@
-// eslint-disable-next-line no-restricted-imports -- TODO: import from specific file
-import { useConfirmationAlert } from ".";
 import { useTagUserStore } from "../stores/user-tags";
 import z from "zod";
+import { useConfirmationAlert } from "./alerts";
 
 export function useTagUser() {
   const setUserTag = useTagUserStore((s) => s.setUserTag);
